% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Age_Donor}
\alias{Age_Donor}
\title{predicted age for each individual}
\usage{
Age_Donor(predict_res)
}
\arguments{
\item{predict_res}{the output from AgingClockCalculator() function}
}
\value{
Return predicted age for each individual
}
\description{
predicted age for each individual
}
