% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AgingClockCalculator}
\alias{AgingClockCalculator}
\title{predict age for each individual}
\usage{
AgingClockCalculator(preprocessed_df, model, marker_gene)
}
\arguments{
\item{preprocessed_df}{the output from PreProcess() function.}

\item{model}{the cell type aging clock model}

\item{marker_gene}{selected marker genes from the corresponding model}
}
\value{
Return dataframe of predicted age
}
\description{
predict age for each individual
}
