% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pre_fun}
\alias{pre_fun}
\title{age prediction}
\usage{
pre_fun(model, selected, final_mtx, preprocessed_df)
}
\arguments{
\item{model}{the cell type aging clock model}

\item{selected}{selected marker genes from the corresponding model}

\item{final_mtx}{input to the model}

\item{preprocessed_df}{the output from PreProcess() function.}
}
\value{
Return dataframe of predicted age
}
\description{
age prediction
}
