% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{PreProcess}
\alias{PreProcess}
\title{Pre-Processing of input dataset}
\usage{
PreProcess(input, cell_type, model, marker_gene)
}
\arguments{
\item{seurat_object}{seurat object with metadata "donor_id", "age"}
}
\value{
Return pre-processed input
}
\description{
input is a seurat object. meta.data should have "donor_id" and "age" columns.
}
