% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pseudocell}
\alias{pseudocell}
\title{Pre-Processing of input dataset}
\usage{
pseudocell(input, size = 15, n = 100, replace = "dynamic")
}
\arguments{
\item{input}{the output from preprocessing() function.}

\item{size}{how many cells would be randomly selected to generate pseudocells}

\item{n}{how many times to repeat random selection}
}
\value{
Return ranked data.frame based on "value"
}
\description{
input could be interested gene list. Each row is one gene and columns are "gene" and "value".
"value" could be log2FC or specific value.
colnames of other metadata should be "subtype".
}
